+++ crates/rover-client/src/error.rs
@@ -1,9 +1,11 @@
-use crate::shared::{CheckTaskStatus, CheckWorkflowResponse, GraphRef, LintResponse};
+use std::fmt::Debug;
+
 use apollo_federation_types::rover::BuildErrors;
 use itertools::Itertools;
 use rover_graphql::GraphQLServiceError;
-use std::fmt::Debug;
 use thiserror::Error;
+
+use crate::shared::{CheckTaskStatus, CheckWorkflowResponse, GraphRef, LintResponse};
 
 /// RoverClientError represents all possible failures that can occur during a client request.
 #[derive(Error, Debug)]
